backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
pydantic
python-dotenv
pyproj==3.7.2
python-multipart-0.0.20
polars-1.33.1